DAD 220 Module 3 Major Activity
The objective of this activity was to create a database for Quantigration, using the company’s return merchandise authorization (RMA) entity relationship diagram (ERD). The ERD is the blueprint for building my relational database. I am given some of the code needed to create the database and had to fill in the rest using what I learned in previous assignments in the course.